change addSeconds on IntDate to take a long rather than an int

+++ src/main/java/org/jose4j/jwt/IntDate.java
@@ -45,7 +45,7 @@ public class IntDate
         return fromSeconds(millisecondsFromEpoch / CONVERSION);
     }
 
-    public void addSeconds(int seconds)
+    public void addSeconds(long seconds)
     {
         value += seconds;
     }
